Move all lit components to dedicated folder.
