+++ src/components/exceptions-table/table-styles.css.ts
