+++ src/components/exceptions-table/utils.ts
